1. organized repo (put all simulations in envs folder) and added tg_ppo algorithm from co_pg 2. added bi-directional wrapper for cvrp
