+++ src/envs/cvrp/cvrp_experimentation/saved_problems/static/mixture_gaussian_20/static_mixture_gaussian_20-customers.json
