+++ src/envs/cvrp/cvrp_experimentation/saved_problems/static/uniform_100/static_uniform_100-customers.json
