+++ src/envs/cvrp/cvrp_experimentation/saved_problems/static/uniform_20/static_uniform_20-customers.json
