+++ src/envs/cvrp/cvrp_experimentation/saved_problems/static/uniform_50/static_uniform_50-customers.json
